Initialize non-negative standard deviations for Bayesian Layer weights.
Variable constraints are only applied after the first optimization step.  This
can lead to unexpected weight values and potential NaN issues on the first
iteration prior to the first optimization step.  This commit changes the logic
to directly initialize a non-negative standard deviation, rather than an
untransformed representation that is subsequently constrained.

PiperOrigin-RevId: 220668146