Add mixture of logistic layer.
Future work can refactor T2T's discretized mixture of logistics loss to use this layer. In the future, we can always add more complexity by for example: 1. splitting out the layer to be a composition of Logistic and Mixture layers; 2. adding more args. I prefer the simplest implementation which works for our applications.

PiperOrigin-RevId: 221368439