Add positive constraint as a replacement for the std dev softplus constraint.
Due to the fact that variable constraints are applied *after* each optimization
step, the softplus function will cause the standard deviation variable to grow
regardless of the derivatives w.r.t. that variable.  This leads to difficult
optimization.  A standard deviation must be positive, so it could make sense to
simply apply a positivity constraint.

PiperOrigin-RevId: 222923785