Improve the initialization scheme for the Bayesian Layers.
This updates the TrainableNormal initialization to a Normal with mean `0` and
standard deviation `sqrt(1. / fan_in)`, where `fan_in` is the input
dimensionality of the layer. Additionally, a small amount of noise is added to
each to break symmetry.  This also includes the He et al. 2015 variant that
uses the standard deviation `sqrt(2. / fan_in)` for layers that are followed
by ReLU nonlinearities, and the Glorot 2010 variant that uses `sqrt(2. / fan_in
+ fan_out)`, where `fan_out` is the output dimensionality of the layer.

PiperOrigin-RevId: 229414516