While collecting trajectories have access to all history of observations.
This does slow down collect, because now we are calling policy over different shapes each time, i.e. t = 1 to T  batch of observations is shaped (b, t) + OBS and t increases every iteration.

But this sems to be unavoidable, unless we know we just want the last time-step's predictions.

PiperOrigin-RevId: 244204123