Cleaning TRAX layers, step 1: remove unused parts.
We may re-add some of the parts we are removing now, but it will be easier to refactor with less parts.

All models (MLP, ResNet, TransformerLM) work fine, other Transformers need refactoring anyway.

Keeping the slax Shared layer so as to use it as the base layer later
(we will want to enable sharing weights by object as that's very natural).

PiperOrigin-RevId: 244427175